update merced model with preliminary guide curve-based costs
